moved page.publications from server folder
