+++ imports/api/pages/pages.publications.js
@@ -1,6 +1,6 @@
 import { Meteor } from 'meteor/meteor';
-import Pages from '../page';
-import { getHost } from '../../@/shared';
+import Pages from './page';
+import { getHost } from '../@/shared';
 
 Meteor.publish('pages', function () {
   const host = getHost(this);
