+++ imports/startup/server/api.js
@@ -13,7 +13,7 @@ import '../../api/@users/user.publications';
 
 import '../../api/pages/page';
 import '../../api/pages/pages.methods';
-import '../../api/pages/server/pages.publications';
+import '../../api/pages/pages.publications';
 
 import '../../api/categories/category';
 import '../../api/categories/categories.methods';
@@ -24,15 +24,11 @@ import '../../api/works/work.publications';
 
 
 /* old */
-// import '../../api/@/methods/accounts-general';
 import '../../api/@/methods/activities-methods';
-// import '../../api/@users/user.admin';
 import '../../api/@/methods/chat-methods';
 import '../../api/@/methods/document-methods';
 import '../../api/@/methods/process-methods';
-// import '../../api/@/methods/profile-methods';
 import '../../api/@/methods/resources-methods';
-// import '../../api/@/methods/work-methods';
 /* old */
